+++ lib/Conversion/StableHLOToTTIR/CMakeLists.txt
@@ -23,4 +23,5 @@ add_mlir_conversion_library(TTMLIRStableHLOToTTIR
   MLIRIR
   MLIRPass
   StablehloPasses
+  TTMLIRTTIRUtils
 )
