+++ lib/Conversion/StableHLOToTTIR/ShardyToTTIRPatterns.cpp
@@ -5,12 +5,10 @@
 #include "ttmlir/Conversion/StableHLOToTTIR/ShardyToTTIR.h"
 
 #include "ttmlir/Conversion/StableHLOToTTIR/ShardingUtils.h"
-#include "ttmlir/Dialect/TT/IR/TT.h"
 #include "ttmlir/Dialect/TT/IR/TTOpsTypes.h"
 #include "ttmlir/Dialect/TT/Utils/Mesh.h"
-#include "ttmlir/Dialect/TTIR/IR/TTIR.h"
 #include "ttmlir/Dialect/TTIR/IR/TTIROps.h"
-#include "ttmlir/Utils.h"
+#include "ttmlir/Dialect/TTIR/Utils/Utils.h"
 
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
@@ -259,7 +257,7 @@ public:
             getTypeConverter()->convertType(localArgType));
 
         auto meshShardOp =
-            ttmlir::utils::createDPSOp<mlir::tt::ttir::MeshShardOp>(
+            mlir::tt::ttir::utils::createDPSOp<mlir::tt::ttir::MeshShardOp>(
                 rewriter, loc, outputType, globalOperand,
                 meshSharding.getShardType(), meshSharding.getShardDirection(),
                 meshSharding.getShardShape(), meshSharding.getShardDims());
@@ -307,7 +305,7 @@ public:
             getTypeConverter()->convertType(opResult.getType()));
 
         auto meshShardOp =
-            ttmlir::utils::createDPSOp<mlir::tt::ttir::MeshShardOp>(
+            mlir::tt::ttir::utils::createDPSOp<mlir::tt::ttir::MeshShardOp>(
                 rewriter, loc, outputType, inputOperand,
                 meshSharding.getShardType(), meshSharding.getShardDirection(),
                 meshSharding.getShardShape(), meshSharding.getShardDims());
