+++ lib/Dialect/TTIR/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(IR)
 add_subdirectory(Pipelines)
 add_subdirectory(Transforms)
+add_subdirectory(Utils)
