+++ lib/Dialect/TTIR/Utils/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_mlir_dialect_library(TTMLIRTTIRUtils
+  Utils.cpp
+)
